requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
folium==0.16.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
mdurl==0.1.2
numpy==1.24.4
packaging==24.0
pandas==1.5.3
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.3.0
scipy==1.10.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-calendar==1.2.0
streamlit_folium==0.20.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
xyzservices==2024.4.0
